move Date component to Utils component folder, initialise Sorter component
